refactor: `ConsoleMenu`로부터 바우처 서비스 메뉴 분리
